A few things
- Removed the voice control. It wasn't properly added as part of the
init chain, so it'd fuck with forked versions who mod stuff. Also it
seems to not work well / was extremely niche. If you want to re-add it,
do so in a fork.

- Changed 5% steppings of speed into a slider. There is currently an
audio bug that occurs when using the offthread CPU option in conjunction
with setting the speed ridiculously low. That's an open and known issue
that only is rearing its head as of late.